Backend/requirements.txt
fastapi
uvicorn
python-multipart
pydantic
pytesseract
opencv-python
pillow
numpy
